Reduce AST size for better parser stack usage
This change updates the types used by the AST in order to minimize stack usage. As a result, more indirection via `Box` and heap allocation is used, and the resulting code actually runs slightly faster.
This fixes #265.

+++ compiler/qsc_frontend/src/typeck/tests.rs
@@ -48,7 +48,8 @@ fn check(source: &str, entry_expr: &str, expect: &Expect) {
     let mut actual = String::new();
     for (id, span, ty) in tys.tys {
         let source = unit.sources.find_offset(span.lo);
-        let code = &source.contents[span.lo - source.offset..span.hi - source.offset];
+        let code = &source.contents
+            [((span.lo - source.offset) as usize)..((span.hi - source.offset) as usize)];
         writeln!(actual, "#{id} {}-{} {code:?} : {ty}", span.lo, span.hi)
             .expect("writing type to string should succeed");
     }
